List events from IPB Calendar in order by time
Previously, events were listed in the order the API returned them, meaning
that the first event in the list could have been the last of the day.  Now
they are in order by time, so the first event in the list should be the
first event of the day, and the last being the last event of the day.