Fix GuildCreateMemberReviewer to submit all users
D4J seemed to change behavior of getPresence when called in the
GuildCreateEvent, as it only returns a value if the user is actually
online.  This caused the list of submissions to the database to only
have the currently online users updated, rather than all users.  Updated
the logic to add all users to the list prior to running the code that
checks presence, since the object in the list will still be updated as
its the same object.  If the change in behavior is reverted in D4J, the
new logic will still function correctly.